init methods moved to corresponding factories
