+++ JDI.Light/JDI.Light/Factories/UIElementFactory.cs
@@ -1,5 +1,14 @@
 ﻿using System;
+using System.Reflection;
+using JDI.Light.Attributes;
 using JDI.Light.Elements.Base;
+using JDI.Light.Elements.Common;
+using JDI.Light.Elements.Composite;
+using JDI.Light.Extensions;
+using JDI.Light.Interfaces.Base;
+using JDI.Light.Interfaces.Common;
+using JDI.Light.Settings;
+using JDI.Light.Utils;
 using OpenQA.Selenium;
 
 namespace JDI.Light.Factories
@@ -28,6 +37,32 @@ namespace JDI.Light.Factories
                 }
             }
             throw new MissingMethodException($"Can't find correct constructor to create instance of type {t}");
+        }
+        
+        public static IBaseElement GetInstanceElement(IBaseElement parent, MemberInfo member)
+        {
+            var type = member.GetMemberType();
+            var instance = (IBaseUIElement)member.GetMemberValue(parent);
+            type = type.IsInterface ? MapInterfaceToElement.ClassFromInterface(type) : type;
+            var element = (UIElement)instance ?? UIElementFactory.CreateInstance(type, member.GetFindsBy());
+            element.Parent = parent;
+            var checkedAttr = member.GetCustomAttribute<IsCheckedAttribute>(false);
+            if (checkedAttr != null && typeof(ICheckBox).IsAssignableFrom(member.GetMemberType()))
+            {
+                var checkBox = (CheckBox)element;
+                checkBox.SetIsCheckedFunc(checkedAttr.CheckedDelegate);
+            }
+            if (parent == null || type != null)
+            {
+                var frameBy = member.GetCustomAttribute<FrameAttribute>(false)?.FrameLocator;
+                if (frameBy != null)
+                    element.FrameLocator = frameBy;
+                By template;
+                if (element.Parent is Form<IConvertible> form && !element.HasLocator
+                                                              && (template = form.LocatorTemplate) != null)
+                    element.Locator = template.FillByTemplate(member.Name);
+            }
+            return element;
         }
     }
 }
